Data - Generates ratio of tree domain variables.

Generates ratio of tree domain summaries from FIESTA::datSumTreedom().

datSumTreeDomRatio(
  ndat,
  ddat,
  uniqueid = "PLT_CN",
  nprefix = NULL,
  dprefix = NULL,
  rprefix = "r",
  datround = NULL,
  savedata = FALSE,
  savedata_opts = NULL,
  gui = FALSE
)

Arguments

ndat

Dataframe, comma-delimited file (*.csv), or shapefile (*.shp). Table from FIESTA::datSumTreeDomRatio() with numerator tree domain variables.

ddat

Dataframe, comma-delimited file (*.csv), or shapefile (*.shp). Table from FIESTA::datSumTreeDomRatio() with numerator tree domain variables.

uniqueid

String. The unique identifier of both tables (default = "PLT_CN").

nprefix

String. The prefix variable identifier of numerator domain variables in ndat.

dprefix

String. The prefix variable identifier of denominator domain variables in ddat.

rprefix

String. The prefix variable identifier of new ratio variables (default="r").

datround

Integer. Number of digits to round ratio values to.

savedata

Logical. If TRUE, saves data to outfolder.

savedata_opts

List. See help(savedata_options()) for a list of options. Only used when savedata = TRUE. If out_layer = NULL, default = 'tsumrat'.

gui

Logical. If gui, user is prompted for parameters.

Value

rdat

Data frame. Table with ratio values (ndat / ddat).

rvars

String vector. Variable names in rdat.

If savedata=TRUE, the data table will be saved to the outfolder:

Author

Tracey S. Frescino

Examples

# Ratio of live and dead basal area
# Live basal area
live <- datSumTreeDom(tree = FIESTA::WYtree, 
                      tuniqueid = "PLT_CN", 
                      tsumvar = "BA", 
                      tfilter = "STATUSCD==1")
#> Warning: COND_STATUS_CD not in table, assuming forested plots with no ACI plots
#> filter removed 5537 records: STATUSCD==1
                      
# Dead basal area                       
dead <- datSumTreeDom(tree = FIESTA::WYtree, 
                      tuniqueid = "PLT_CN", 
                      tsumvar = "BA", 
                      tfilter = "STATUSCD == 2 & STANDING_DEAD_CD == 1")
#> Warning: COND_STATUS_CD not in table, assuming forested plots with no ACI plots
#> filter removed 13474 records: STATUSCD == 2 & STANDING_DEAD_CD == 1

# Ratio of dead and live basal area                        
ratio <- FIESTA:::datSumTreeDomRatio(ndat = dead$tdomdat,
                                     ddat = live$tdomdat)    

str(ratio, max.level = 1)
#> List of 2
#>  $ rdat :'data.frame':	409 obs. of  20 variables:
#>  $ rvars: chr [1:19] "r19" "r65" "r66" "r93" ...

head(ratio$rdat)         
#>           PLT_CN r19 r65       r66 r93 r96 r101 r106 r108      r113       r122
#> 1 40404729010690   0   0 0.0000000   0   0    0    0    0 0.4699766 0.00000000
#> 2 40404730010690   0   0 0.0000000   0   0    0    0    0 0.4070551 0.05778194
#> 3 40404738010690   0   0 0.0000000   0   0    0    0    0 1.0545870        Inf
#> 4 40404743010690   0   0 0.0000000   0   0    0    0    0 0.0000000 0.07211513
#> 5 40404758010690   0   0 0.0000000   0   0    0    0    0 0.0000000 0.22292733
#> 6 40404762010690   0   0 0.1842018   0   0    0    0    0 0.0000000        Inf
#>   r202 r313 r375 r475 r544 r745 r746 r749 r823
#> 1    0    0    0    0    0    0    0    0    0
#> 2    0    0    0    0    0    0    0    0    0
#> 3    0    0    0    0    0    0    0    0    0
#> 4    0    0    0    0    0    0    0    0    0
#> 5    0    0    0    0    0    0    0    0    0
#> 6    0    0    0    0    0    0    0    0    0