Model-Assisted module - Generate model-assisted area estimates.

Generates area estimates by estimation unit. Estimates are calculated from McConville et al. (2018)'s mase R package.

modMAarea(
  MApopdat,
  MAmethod,
  FIA = TRUE,
  prednames = NULL,
  modelselect = FALSE,
  landarea = "FOREST",
  pcfilter = NULL,
  rowvar = NULL,
  colvar = NULL,
  bootstrap = FALSE,
  returntitle = FALSE,
  savedata = FALSE,
  table_opts = NULL,
  title_opts = NULL,
  savedata_opts = NULL,
  gui = FALSE,
  modelselect_bydomain = FALSE,
  ...
)

Arguments

MApopdat

List. Population data objects returned from modMApop().

MAmethod

String. mase (i.e., model-assisted) method to use ('greg', 'gregEN', 'ratio').

FIA

Logical. If TRUE, the finite population term is removed from estimator to match FIA estimates.

prednames

String vector. Name(s) of predictor variables to include in model.

modelselect

Logical. If TRUE, an elastic net regression model is fit to the entire plot level data, and the variables selected in that model are used for the proceeding estimation.

landarea

String. The sample area filter for estimates ('ALL', 'FOREST', 'TIMBERLAND'). If landarea=FOREST, filtered to COND_STATUS_CD = 1; If landarea=TIMBERLAND, filtered to SITECLCD in(1:6) and RESERVCD = 0.

pcfilter

String. A filter for plot or cond attributes (including pltassgn). Must be R logical syntax.

rowvar

String. Name of the row domain variable in cond or tree. If only one domain, rowvar = domain variable. If more than one domain, include colvar. If no domain, rowvar = NULL.

colvar

String. Name of the column domain variable in cond or tree.

bootstrap

Logical. If TRUE, returns bootstrap variance estimates, otherwise uses Horvitz-Thompson estimator under simple random sampling without replacement.

returntitle

Logical. If TRUE, returns title(s) of the estimation table(s).

savedata

Logical. If TRUE, saves table(s) to outfolder.

table_opts

List. See help(table_options()) for a list of options.

title_opts

List. See help(title_options()) for a list of options.

savedata_opts

List. See help(savedata_options()) for a list of options. Only used when savedata = TRUE.

gui

Logical. If gui, user is prompted for parameters.

modelselect_bydomain

Logical. If TRUE, modelselection will occur at the domain level as specified by rowvar and/or colvar and not at the level of the entire sample.

...

Parameters for modMApop() if MApopdat is NULL.

Value

If FIA=TRUE or unitvar=NULL and colvar=NULL, one data frame is returned with tree estimates and percent sample errors. Otherwise, a list is returned with tree estimates in one data frame (est) and percent sample errors in another data frame (est.pse). If rawdata=TRUE, another list is returned including raw data used in the estimation process. If addtitle=TRUE and returntitle=TRUE, the title for est/pse is returned. If savedata=TRUE, all data frames are written to outfolder.

est

Data frame. Tree estimates by rowvar, colvar (and estimation unit). If FIA=TRUE or one estimation unit and colvar=NULL, estimates and percent sampling error are in one data frame.

pse

Data frame. Percent sampling errors for estimates by rowvar and colvar (and estimation unit).

titlelst

List with 1 or 2 string vectors. If returntitle=TRUE a list with table title(s). The list contains one title if est and pse are in the same table and two titles if est and pse are in separate tables.

raw

List of data frames. If rawdata=TRUE, a list including: number of plots by plot status, if in dataset (plotsampcnt); number of conditions by condition status (condsampcnt); data used for post-stratification (stratdat); and 1-8 tables with calculated variables used for processing estimates and percent sampling error for table cell values and totals (See processing data below).

Raw data

plotsampcnt

Table. Number of plots by plot status (ex. sampled forest on plot, sampled nonforest, nonsampled).

condsampcnt

DF. Number of conditions by condition status (forest land, nonforest land, noncensus water, census water, nonsampled).

stratdat

Data frame. Strata information by estimation unit.

Variable	Description

ESTUNIT	estimation unit

STRATA	strata

ACRES	area by strata for estimation unit

n.strata	number of plots in strata (and estimation unit)

n.total	number of plots for estimation unit

TOTACRES	total area for estimation unit

strwt	proportion of area (or number of plots) by strata (strata weight)

expfac.strata	expansion factor (in area unit (e.g., acres) by strata (areavar/n.strata)

processing data

Data frames. Separate data frames containing calculated variables used in estimation process. The number of processing tables depends on the input parameters. The tables include: total by estimation unit (unit.totest); rowvar totals (unit.rowest), and if colvar is not NULL, colvar totals, (unit.colvar); and a combination of rowvar and colvar (unit.grpvar). If FIA=TRUE, the raw data for the summed estimation units are also included (totest, rowest, colest, grpest, respectively). These tables do not included estimate proportions (nhat and nhat.var).

The data frames include the following information:

Variable	Description

nhat	estimated proportion of trees

nhat.var	estimated variance of estimated proportion of trees

ACRES	total area for estimation unit

est	estimated area of trees nhat*ACRES

est.var	estimated variance of estimated area of trees nhat.var*areavar^2

est.se	standard error of estimated area of trees sqrt(est.var)

est.cv	coefficient of variation of estimated area of trees est.se/est

pse	percent sampling error of estimate est.cv*100

CI99left	left tail of 99 percent confidence interval for estimated area

CI99right	right tail of 99 percent confidence interval for estimated area

CI95left	left tail of 95 percent confidence interval for estimated area

CI95right	right tail of 95 percent confidence interval for estimated area

CI67left	left tail of 67 percent confidence interval for estimated area

CI67right	right tail of 67 percent confidence interval for estimated area

Table(s) are also written to outfolder.

Details

If variables are NULL, then it will prompt user to input variables.

Necessary variables:

Data	Variable	Description

tree	tuniqueid	Unique identifier for each plot, to link to pltstrat (ex. PLT_CN).

CONDID	Unique identifier of each condition on plot, to link to cond. Set CONDID=1, if only 1 condition per plot.

cond	cuniqueid	Unique identifier for each plot, to link to pltstrat (ex. PLT_CN).

CONDID	Unique identifier of each condition on plot. Set CONDID=1, if only 1 condition per plot.

CONDPROP_UNADJ	Unadjusted proportion of condition on each plot. Set CONDPROP_UNADJ=1, if only 1 condition per plot.

COND_STATUS_CD	Status of each forested condition on plot (i.e. accessible forest, nonforest, water, etc.)

NF_COND_STATUS_CD	If ACI=TRUE. Status of each nonforest condition on plot (i.e. accessible nonforest, nonsampled nonforest)

SITECLCD	If landarea=TIMBERLAND. Measure of site productivity.

RESERVCD	If landarea=TIMBERLAND. Reserved status.

pltstrat	puniqueid	Unique identifier for each plot, to link to cond (ex. CN).

STATECD	Identifies state each plot is located in.

INVYR	Identifies inventory year of each plot.

PLOT_STATUS_CD	Status of each plot (i.e. sampled, nonsampled). If not included, all plots are assumed as sampled.

Reference names are available for the following variables:
ADFORCD, AGENTCD, CCLCD, DECAYCD, DSTRBCD, KINDCD, OWNCD, OWNGRPCD, FORTYPCD, FLDTYPCD, FORTYPCDCALC, TYPGRPCD, FORINDCD, RESERVCD, LANDCLCD, STDSZCD, FLDSZCD, PHYSCLCD, MIST_CL_CD, PLOT_STATUS_CD, STATECD, TREECLCD, TRTCD, SPCD, SPGRPCD

Note

ADJUSTMENT FACTOR:
The adjustment factor is necessary to account for nonsampled conditions. It is calculated for each estimation unit by strata. by summing the unadjusted proportions of the subplot, microplot, and macroplot (i.e. *PROP_UNADJ) and dividing by the number of plots in the strata/estimation unit).

An adjustment factor is determined for each tree based on the size of the plot it was measured on. This is identified using TPA_UNADJ as follows:

PLOT SIZE	TPA_UNADJ

SUBPLOT	6.018046

MICROPLOT	74.965282

MACROPLOT	0.999188

If ACI=FALSE, only nonsampled forest conditions are accounted for in the adjustment factor.
If ACI=TRUE, the nonsampled nonforest conditions are removed as well and accounted for in adjustment factor. This is if you are interested in estimates for all lands or nonforest lands in the All-Condition-Inventory.

stratcombine:
If MAmethod='PS', and stratcombine=TRUE, and less than 2 plots in any one estimation unit, all estimation units with 10 or less plots are combined. The current method for combining is to group the estimation unit with less than 10 plots with the estimation unit following in consecutive order (numeric or alphabetical), restrained by survey unit (UNITCD) if included in dataset, and continuing until the number of plots equals 10. If there are no estimation units following in order, it is combined with the estimation unit previous in order.

autoxreduce:
If MAmethod='GREG', and autoxreduce=TRUE, and there is an error because of multicolinearity, a variable reduction method is applied to remove correlated variables. The method used is based on the variance-inflation factor (vif) from a linear model. The vif estimates how much the variance of each x variable is inflated due to mulitcolinearity in the model.

rowlut/collut:
There are several objectives for including rowlut/collut look-up tables: 1) to include descriptive names that match row/column codes in the input table; 2) to use number codes that match row/column names in the input table for ordering rows; 3) to add rows and/or columns with 0 values for consistency. No duplicate names are allowed.

Include 2 columns in the table:
1-the merging variable with same name as the variable in the input merge table;
2-the ordering or descriptive variable.
If the ordering variable is the rowvar/colvar in the input table and the descriptive variable is in rowlut/collut, set row.orderby/col.orderby equal to rowvar/colvar. If the descriptive variable is the rowvar/colvar in the input table, and the ordering code variable is in rowlut/collut, set row.orderby/col.orderby equal to the variable name of the code variable in rowlut/collut.

UNITS:
The following variables are converted from pounds (from FIA database) to short tons by multiplying the variable by 0.0005. DRYBIO_AG, DRYBIO_BG, DRYBIO_WDLD_SPP, DRYBIO_SAPLING, DRYBIO_STUMP, DRYBIO_TOP, DRYBIO_BOLE, DRYBIOT, DRYBIOM, DRYBIOTB, JBIOTOT, CARBON_BG, CARBON_AG

MORTALITY:
For Interior-West FIA, mortality estimates are mainly based on whether a tree has died within the last 5 years of when the plot was measured. If a plot was remeasured, mortality includes trees that were alive the previous visit but were dead in the next visit. If a tree was standing the previous visit, but was not standing in the next visit, no diameter was collected (DIA = NA) but the tree is defined as mortality.

Common tree filters:

FILTER	DESCRIPTION

"STATUSCD == 1"	Live trees

"STATUSCD == 2"	Dead trees

"TPAMORT_UNADJ > 0"	Mortality trees

"STATUSCD == 2 & DIA >= 5.0"	Dead trees >= 5.0 inches diameter

"STATUSCD == 2 & AGENTCD == 30"	Dead trees from fire

References

Kelly McConville, Becky Tang, George Zhu, Shirley Cheung, and Sida Li (2018). mase: Model-Assisted Survey Estimation. R package version 0.1.2 https://cran.r-project.org/package=mase

Author

Tracey S. Frescino

Examples

# \donttest{
# Set up population dataset (see ?modMApop() for more information)
MApopdat <- modMApop(popTabs = list(tree = FIESTA::WYtree,
                                    cond = FIESTA::WYcond),
                     pltassgn = FIESTA::WYpltassgn,
                     pltassgnid = "CN",
                     unitarea = FIESTA::WYunitarea,
                     unitvar = "ESTN_UNIT",
                     unitzonal = FIESTA::WYunitzonal,
                     prednames = c("dem", "tcc", "tpi", "tnt"),
                     predfac = "tnt")
#> For FIA estimation, adjustment factors are calculated to account for plots with partial nonresponse.
#> ...there are 14 nonsampled forest conditions in the dataset.
#> COND_STATUS_CD != 5
#> filter removed 14 records: COND_STATUS_CD != 5

# Use GREG estimator to estimate area of forest land in our population
mod1 <- modMAarea(MApopdat = MApopdat, 
          MAmethod = "greg", 
          landarea = "FOREST")
#> COND_STATUS_CD == 1
#> filter removed 2620 records: COND_STATUS_CD == 1
#> COND_STATUS_CD == 1
#> generating estimates using mase::greg function...
#> using the following predictors...dem, tcc, tpi, tnt2
#> getting output...

str(mod1)
#> List of 4
#>  $ est    :Classes ‘data.table’ and 'data.frame':	23 obs. of  3 variables:
#>   ..$ ESTN_UNIT             : chr [1:23] "1" "11" "13" "15" ...
#>   ..$ Estimate              : num [1:23] 574512 610054 1009706 34921 207027 ...
#>   ..$ Percent Sampling Error: num [1:23] 10.66 9.91 8.45 53.02 19.97 ...
#>   ..- attr(*, ".internal.selfref")=<externalptr> 
#>   ..- attr(*, "sorted")= chr "ESTN_UNIT"
#>  $ raw    :List of 9
#>   ..$ unit_totest  :'data.frame':	23 obs. of  18 variables:
#>   .. ..$ ESTN_UNIT : chr [1:23] "1" "11" "13" "15" ...
#>   .. ..$ nhat      : num [1:23] 0.2083 0.3321 0.1703 0.0244 0.1612 ...
#>   .. ..$ nhat.var  : num [1:23] 0.000493 0.001082 0.000207 0.000168 0.001037 ...
#>   .. ..$ NBRPLT    : int [1:23] 133 85 290 70 58 128 86 132 175 79 ...
#>   .. ..$ NBRPLT.gt0: int [1:23] 24 26 53 2 8 22 2 38 7 2 ...
#>   .. ..$ ACRES     : num [1:23] 2757613 1837124 5930088 1428579 1283969 ...
#>   .. ..$ AREAUSED  : num [1:23] 2757613 1837124 5930088 1428579 1283969 ...
#>   .. ..$ est       : num [1:23] 574512 610054 1009706 34921 207027 ...
#>   .. ..$ est.var   : num [1:23] 3.75e+09 3.65e+09 7.27e+09 3.43e+08 1.71e+09 ...
#>   .. ..$ est.se    : num [1:23] 61249 60429 85278 18513 41341 ...
#>   .. ..$ est.cv    : num [1:23] 0.1066 0.0991 0.0845 0.5302 0.1997 ...
#>   .. ..$ pse       : num [1:23] 10.66 9.91 8.45 53.02 19.97 ...
#>   .. ..$ CI99left  : num [1:23] 416744 454398 790045 0 100540 ...
#>   .. ..$ CI99right : num [1:23] 732280 765709 1229368 82608 313514 ...
#>   .. ..$ CI95left  : num [1:23] 454466 491615 842564 0 126000 ...
#>   .. ..$ CI95right : num [1:23] 694559 728493 1176848 71206 288053 ...
#>   .. ..$ CI68left  : num [1:23] 513602 549959 924901 16510 165915 ...
#>   .. ..$ CI68right : num [1:23] 635422 670148 1094512 53331 248139 ...
#>   ..$ domdat       :'data.frame':	3210 obs. of  19 variables:
#>   .. ..$ PLT_CN         : chr [1:3210] "40404728010690" "40404729010690" "40404730010690" "40404731010690" ...
#>   .. ..$ CONDID         : num [1:3210] 1 1 1 1 1 1 1 1 1 2 ...
#>   .. ..$ CONDPROP_UNADJ : num [1:3210] 1 1 1 1 1 1 1 1 0.75 0.25 ...
#>   .. ..$ SUBPPROP_UNADJ : num [1:3210] 1 1 1 1 1 1 1 1 0.75 0.25 ...
#>   .. ..$ MACRPROP_UNADJ : logi [1:3210] NA NA NA NA NA NA ...
#>   .. ..$ MICRPROP_UNADJ : num [1:3210] 1 1 1 1 1 1 1 1 0.75 0.25 ...
#>   .. ..$ ESTN_UNIT      : Factor w/ 23 levels "1","3","5","7",..: 1 1 1 1 1 1 1 1 1 1 ...
#>   .. ..$ dem            : int [1:3210] 2372 2382 2274 2203 2276 2178 2214 2229 2576 2576 ...
#>   .. ..$ tcc            : int [1:3210] 0 5 45 0 0 0 0 0 16 16 ...
#>   .. ..$ tpi            : num [1:3210] 10.625 0.75 -12.125 -0.125 1.25 ...
#>   .. ..$ tnt2           : num [1:3210] 1 1 0 1 1 1 1 1 0 0 ...
#>   .. ..$ cadjfac        : num [1:3210] 1 1 1 1 1 1 1 1 1 1 ...
#>   .. ..$ ADJ_FACTOR_SUBP: num [1:3210] 1 1 1 1 1 1 1 1 1 1 ...
#>   .. ..$ ADJ_FACTOR_MACR: num [1:3210] 0 0 0 0 0 0 0 0 0 0 ...
#>   .. ..$ ADJ_FACTOR_MICR: num [1:3210] 1 1 1 1 1 1 1 1 1 1 ...
#>   .. ..$ CONDPROP_ADJ   : num [1:3210] 1 1 1 1 1 1 1 1 0.75 0.25 ...
#>   .. ..$ COND_STATUS_CD : int [1:3210] NA 1 1 NA NA NA NA NA NA 1 ...
#>   .. ..$ TOTAL          : num [1:3210] NA 1 1 NA NA NA NA NA NA 1 ...
#>   .. ..$ AREA_ADJ       : num [1:3210] 0 1 1 0 0 0 0 0 0 0.25 ...
#>   ..$ module       : chr "MA"
#>   ..$ esttype      : chr "AREA"
#>   ..$ MAmethod     : chr "greg"
#>   ..$ predselectlst:List of 1
#>   .. ..$ totest:Classes ‘data.table’ and 'data.frame':	23 obs. of  6 variables:
#>   .. .. ..$ ESTN_UNIT: Factor w/ 23 levels "1","3","5","7",..: 1 2 3 4 5 6 7 8 9 10 ...
#>   .. .. ..$ TOTAL    : num [1:23] 1 1 1 1 1 1 1 1 1 1 ...
#>   .. .. ..$ dem      : num [1:23] -7.18e-05 1.38e-04 -1.44e-04 1.96e-04 2.29e-04 ...
#>   .. .. ..$ tcc      : num [1:23] 0.01676 0.00724 0.01533 0.00902 0.01985 ...
#>   .. .. ..$ tpi      : num [1:23] 2.44e-03 -8.37e-05 -1.17e-02 -3.78e-03 -4.83e-03 ...
#>   .. .. ..$ tnt2     : num [1:23] -0.26 -0.164 0.306 -0.376 -0.069 ...
#>   .. .. ..- attr(*, ".internal.selfref")=<externalptr> 
#>   ..$ rowvar       : chr "TOTAL"
#>   ..$ colvar       : chr "NONE"
#>   ..$ areaunits    : chr "acres"
#>  $ statecd: int 56
#>  $ invyr  : int [1:3] 2011 2012 2013
          
# Use GREG estimator to estimate area of forest land by forest type and
# stand-size class
mod2 <- modMAarea(MApopdat = MApopdat,
          MAmethod = "greg",
          landarea = "FOREST",
          rowvar = "FORTYPCD",
          colvar = "STDSZCD")
#> COND_STATUS_CD == 1
#> filter removed 2620 records: COND_STATUS_CD == 1
#> COND_STATUS_CD == 1
#> generating estimates using mase::greg function...
#> using the following predictors...dem, tcc, tpi, tnt2
#> getting output...
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
          
str(mod2)
#> List of 5
#>  $ est    :Classes ‘data.table’ and 'data.frame':	437 obs. of  7 variables:
#>   ..$ ESTN_UNIT  : chr [1:437] "1" "1" "1" "1" ...
#>   ..$ Forest type: chr [1:437] "182" "184" "185" "201" ...
#>   ..$ 1          : chr [1:437] "220976.1" "--" "--" "37185.9" ...
#>   ..$ 2          : chr [1:437] "--" "--" "--" "--" ...
#>   ..$ 3          : chr [1:437] "254987.7" "--" "--" "--" ...
#>   ..$ 5          : chr [1:437] "--" "--" "--" "--" ...
#>   ..$ Total      : chr [1:437] "475963.8" "--" "--" "37185.9" ...
#>   ..- attr(*, ".internal.selfref")=<externalptr> 
#>  $ pse    :Classes ‘data.table’ and 'data.frame':	437 obs. of  7 variables:
#>   ..$ ESTN_UNIT  : chr [1:437] "1" "1" "1" "1" ...
#>   ..$ Forest type: chr [1:437] "182" "184" "185" "201" ...
#>   ..$ 1          : chr [1:437] "49.03" "--" "--" "259.58" ...
#>   ..$ 2          : chr [1:437] "--" "--" "--" "--" ...
#>   ..$ 3          : chr [1:437] "43.02" "--" "--" "--" ...
#>   ..$ 5          : chr [1:437] "--" "--" "--" "--" ...
#>   ..$ Total      : chr [1:437] "30.06" "--" "--" "259.58" ...
#>   ..- attr(*, ".internal.selfref")=<externalptr> 
#>  $ raw    :List of 12
#>   ..$ unit_totest  :'data.frame':	23 obs. of  18 variables:
#>   .. ..$ ESTN_UNIT : chr [1:23] "1" "11" "13" "15" ...
#>   .. ..$ nhat      : num [1:23] 0.2083 0.3321 0.1703 0.0244 0.1612 ...
#>   .. ..$ nhat.var  : num [1:23] 0.000493 0.001082 0.000207 0.000168 0.001037 ...
#>   .. ..$ NBRPLT    : int [1:23] 133 85 290 70 58 128 86 132 175 79 ...
#>   .. ..$ NBRPLT.gt0: int [1:23] 24 26 53 2 8 22 2 38 7 2 ...
#>   .. ..$ ACRES     : num [1:23] 2757613 1837124 5930088 1428579 1283969 ...
#>   .. ..$ AREAUSED  : num [1:23] 2757613 1837124 5930088 1428579 1283969 ...
#>   .. ..$ est       : num [1:23] 574512 610054 1009706 34921 207027 ...
#>   .. ..$ est.var   : num [1:23] 3.75e+09 3.65e+09 7.27e+09 3.43e+08 1.71e+09 ...
#>   .. ..$ est.se    : num [1:23] 61249 60429 85278 18513 41341 ...
#>   .. ..$ est.cv    : num [1:23] 0.1066 0.0991 0.0845 0.5302 0.1997 ...
#>   .. ..$ pse       : num [1:23] 10.66 9.91 8.45 53.02 19.97 ...
#>   .. ..$ CI99left  : num [1:23] 416744 454398 790045 0 100540 ...
#>   .. ..$ CI99right : num [1:23] 732280 765709 1229368 82608 313514 ...
#>   .. ..$ CI95left  : num [1:23] 454466 491615 842564 0 126000 ...
#>   .. ..$ CI95right : num [1:23] 694559 728493 1176848 71206 288053 ...
#>   .. ..$ CI68left  : num [1:23] 513602 549959 924901 16510 165915 ...
#>   .. ..$ CI68right : num [1:23] 635422 670148 1094512 53331 248139 ...
#>   ..$ unit_rowest  :'data.frame':	135 obs. of  19 variables:
#>   .. ..$ ESTN_UNIT  : chr [1:135] "1" "1" "1" "1" ...
#>   .. ..$ Forest type: Factor w/ 18 levels "182","184","185",..: 1 4 5 7 8 10 11 17 18 5 ...
#>   .. ..$ nhat       : num [1:135] 0.1726 0.0135 0.2302 -0.0139 0.0147 ...
#>   .. ..$ nhat.var   : num [1:135] 0.00269 0.00123 0.00309 0.00337 0.00162 ...
#>   .. ..$ NBRPLT     : int [1:135] 24 24 24 24 24 24 24 24 24 26 ...
#>   .. ..$ NBRPLT.gt0 : int [1:135] 2 1 4 4 1 4 2 4 2 18 ...
#>   .. ..$ ACRES      : num [1:135] 2757613 2757613 2757613 2757613 2757613 ...
#>   .. ..$ AREAUSED   : num [1:135] 2757613 2757613 2757613 2757613 2757613 ...
#>   .. ..$ est        : num [1:135] 475964 37186 634707 -38226 40580 ...
#>   .. ..$ est.var    : num [1:135] 2.05e+10 9.32e+09 2.35e+10 2.56e+10 1.23e+10 ...
#>   .. ..$ est.se     : num [1:135] 143066 96527 153302 160065 110859 ...
#>   .. ..$ est.cv     : num [1:135] 0.301 2.596 0.242 -4.187 2.732 ...
#>   .. ..$ pse        : num [1:135] 30.1 259.6 24.2 -418.7 273.2 ...
#>   .. ..$ CI99left   : num [1:135] 107450 0 239828 0 0 ...
#>   .. ..$ CI99right  : num [1:135] 844477 285822 1029586 374074 326134 ...
#>   .. ..$ CI95left   : num [1:135] 195560 0 334241 0 0 ...
#>   .. ..$ CI95right  : num [1:135] 756368 226375 935173 275496 257860 ...
#>   .. ..$ CI68left   : num [1:135] 333691 0 482255 0 0 ...
#>   .. ..$ CI68right  : num [1:135] 618237 133178 787159 120952 150825 ...
#>   ..$ unit_colest  :'data.frame':	92 obs. of  19 variables:
#>   .. ..$ ESTN_UNIT       : chr [1:92] "1" "1" "1" "1" ...
#>   .. ..$ Stand-size class: Factor w/ 4 levels "1","2","3","5": 1 2 3 4 1 2 3 4 1 2 ...
#>   .. ..$ nhat            : num [1:92] 0.40336 -0.00388 0.4542 0.09232 0.62726 ...
#>   .. ..$ nhat.var        : num [1:92] 0.0083 0.00354 0.00554 0.00168 0.007 ...
#>   .. ..$ NBRPLT          : num [1:92] 24 24 24 24 26 26 26 26 53 53 ...
#>   .. ..$ NBRPLT.gt0      : num [1:92] 9 6 7 2 18 4 4 1 23 16 ...
#>   .. ..$ ACRES           : num [1:92] 2757613 2757613 2757613 2757613 1837124 ...
#>   .. ..$ AREAUSED        : num [1:92] 2757613 2757613 2757613 2757613 1837124 ...
#>   .. ..$ est             : num [1:92] 1112321 -10692 1252510 254595 1152362 ...
#>   .. ..$ est.var         : num [1:92] 6.31e+10 2.69e+10 4.21e+10 1.28e+10 2.36e+10 ...
#>   .. ..$ est.se          : num [1:92] 251214 164035 205290 113005 153718 ...
#>   .. ..$ est.cv          : num [1:92] 0.226 -15.342 0.164 0.444 0.133 ...
#>   .. ..$ pse             : num [1:92] 22.6 -1534.2 16.4 44.4 13.3 ...
#>   .. ..$ CI99left        : num [1:92] 465235 0 723719 0 756410 ...
#>   .. ..$ CI99right       : num [1:92] 1759406 411833 1781301 545676 1548314 ...
#>   .. ..$ CI95left        : num [1:92] 619950 0 850149 33109 851079 ...
#>   .. ..$ CI95right       : num [1:92] 1604691 310810 1654870 476080 1453644 ...
#>   .. ..$ CI68left        : num [1:92] 862499 0 1048358 142216 999495 ...
#>   .. ..$ CI68right       : num [1:92] 1362143 152434 1456662 366973 1305228 ...
#>   ..$ unit_grpest  :'data.frame':	217 obs. of  21 variables:
#>   .. ..$ ESTN_UNIT       : chr [1:217] "1" "1" "1" "1" ...
#>   .. ..$ grpvar          : chr [1:217] "182#1" "182#3" "201#1" "221#1" ...
#>   .. ..$ nhat            : num [1:217] 0.08013 0.09247 0.01348 0.23017 -0.00585 ...
#>   .. ..$ nhat.var        : num [1:217] 0.00154 0.00158 0.00123 0.00309 0.00192 ...
#>   .. ..$ NBRPLT          : num [1:217] 24 24 24 24 24 24 24 24 24 24 ...
#>   .. ..$ NBRPLT.gt0      : num [1:217] 1 1 1 4 2 2 1 3 1 1 ...
#>   .. ..$ Forest type     : Factor w/ 18 levels "182","184","185",..: 1 1 4 5 7 7 8 10 10 11 ...
#>   .. ..$ Stand-size class: Factor w/ 4 levels "1","2","3","5": 1 3 1 1 1 2 2 2 3 1 ...
#>   .. ..$ ACRES           : num [1:217] 2757613 2757613 2757613 2757613 2757613 ...
#>   .. ..$ AREAUSED        : num [1:217] 2757613 2757613 2757613 2757613 2757613 ...
#>   .. ..$ est             : num [1:217] 220976 254988 37186 634707 -16120 ...
#>   .. ..$ est.var         : num [1:217] 1.17e+10 1.20e+10 9.32e+09 2.35e+10 1.46e+10 ...
#>   .. ..$ est.se          : num [1:217] 108343 109685 96527 153302 120925 ...
#>   .. ..$ est.cv          : num [1:217] 0.49 0.43 2.596 0.242 -7.502 ...
#>   .. ..$ pse             : num [1:217] 49 43 259.6 24.2 -750.2 ...
#>   .. ..$ CI99left        : num [1:217] 0 0 0 239828 0 ...
#>   .. ..$ CI99right       : num [1:217] 500049 537517 285822 1029586 295363 ...
#>   .. ..$ CI95left        : num [1:217] 8628 40010 0 334241 0 ...
#>   .. ..$ CI95right       : num [1:217] 433324 469966 226375 935173 220889 ...
#>   .. ..$ CI68left        : num [1:217] 113233 145911 0 482255 0 ...
#>   .. ..$ CI68right       : num [1:217] 328719 364064 133178 787159 104135 ...
#>   ..$ domdat       :'data.frame':	590 obs. of  21 variables:
#>   .. ..$ PLT_CN         : chr [1:590] "40404729010690" "40404730010690" "40404737010690" "40404738010690" ...
#>   .. ..$ CONDID         : num [1:590] 1 1 2 1 1 1 2 1 1 1 ...
#>   .. ..$ CONDPROP_UNADJ : num [1:590] 1 1 0.25 1 1 ...
#>   .. ..$ SUBPPROP_UNADJ : num [1:590] 1 1 0.25 1 1 ...
#>   .. ..$ MACRPROP_UNADJ : logi [1:590] NA NA NA NA NA NA ...
#>   .. ..$ MICRPROP_UNADJ : num [1:590] 1 1 0.25 1 1 1 0.5 1 1 1 ...
#>   .. ..$ ESTN_UNIT      : Factor w/ 23 levels "1","3","5","7",..: 1 1 1 1 1 1 1 1 1 1 ...
#>   .. ..$ dem            : int [1:590] 2382 2274 2576 2434 2202 2177 2011 1861 2161 2124 ...
#>   .. ..$ tcc            : int [1:590] 5 45 16 27 1 16 5 19 0 3 ...
#>   .. ..$ tpi            : num [1:590] 0.75 -12.12 2.5 -0.75 -8.62 ...
#>   .. ..$ tnt2           : num [1:590] 1 0 0 0 1 1 1 0 1 1 ...
#>   .. ..$ cadjfac        : num [1:590] 1 1 1 1 1 ...
#>   .. ..$ ADJ_FACTOR_SUBP: num [1:590] 1 1 1 1 1 ...
#>   .. ..$ ADJ_FACTOR_MACR: num [1:590] 0 0 0 0 0 0 0 0 0 0 ...
#>   .. ..$ ADJ_FACTOR_MICR: num [1:590] 1 1 1 1 1 ...
#>   .. ..$ CONDPROP_ADJ   : num [1:590] 1 1 0.25 1 1 ...
#>   .. ..$ COND_STATUS_CD : int [1:590] 1 1 1 1 1 1 1 1 1 1 ...
#>   .. ..$ FORTYPCD       : int [1:590] 366 201 999 366 901 221 221 221 221 999 ...
#>   .. ..$ STDSZCD        : int [1:590] 1 1 5 3 3 1 1 1 1 5 ...
#>   .. ..$ TOTAL          : num [1:590] 1 1 1 1 1 1 1 1 1 1 ...
#>   .. ..$ AREA_ADJ       : num [1:590] 1 1 0.25 1 1 ...
#>   ..$ module       : chr "MA"
#>   ..$ esttype      : chr "AREA"
#>   ..$ MAmethod     : chr "greg"
#>   ..$ predselectlst:List of 2
#>   .. ..$ totest:Classes ‘data.table’ and 'data.frame':	23 obs. of  6 variables:
#>   .. .. ..$ ESTN_UNIT: Factor w/ 23 levels "1","3","5","7",..: 1 2 3 4 5 6 7 8 9 10 ...
#>   .. .. ..$ TOTAL    : num [1:23] 1 1 1 1 1 1 1 1 1 1 ...
#>   .. .. ..$ dem      : num [1:23] -7.18e-05 1.38e-04 -1.44e-04 1.96e-04 2.29e-04 ...
#>   .. .. ..$ tcc      : num [1:23] 0.01676 0.00724 0.01533 0.00902 0.01985 ...
#>   .. .. ..$ tpi      : num [1:23] 2.44e-03 -8.37e-05 -1.17e-02 -3.78e-03 -4.83e-03 ...
#>   .. .. ..$ tnt2     : num [1:23] -0.26 -0.164 0.306 -0.376 -0.069 ...
#>   .. .. ..- attr(*, ".internal.selfref")=<externalptr> 
#>   .. ..$ rowest:Classes ‘data.table’ and 'data.frame':	23 obs. of  6 variables:
#>   .. .. ..$ ESTN_UNIT: Factor w/ 23 levels "1","3","5","7",..: 1 2 3 4 5 6 7 8 9 10 ...
#>   .. .. ..$ TOTAL    : num [1:23] 1 1 1 1 1 1 1 1 1 1 ...
#>   .. .. ..$ dem      : num [1:23] -7.18e-05 1.38e-04 -1.44e-04 1.96e-04 2.29e-04 ...
#>   .. .. ..$ tcc      : num [1:23] 0.01676 0.00724 0.01533 0.00902 0.01985 ...
#>   .. .. ..$ tpi      : num [1:23] 2.44e-03 -8.37e-05 -1.17e-02 -3.78e-03 -4.83e-03 ...
#>   .. .. ..$ tnt2     : num [1:23] -0.26 -0.164 0.306 -0.376 -0.069 ...
#>   .. .. ..- attr(*, ".internal.selfref")=<externalptr> 
#>   ..$ rowvar       : chr "FORTYPCD"
#>   ..$ colvar       : chr "STDSZCD"
#>   ..$ areaunits    : chr "acres"
#>  $ statecd: int 56
#>  $ invyr  : int [1:3] 2011 2012 2013
# }