Small area module - Generate small area tree estimates.

Generates small area estimates by domain and/or tree domain (and estimation unit).

modSAest(
  SApopdatlst = NULL,
  SAdomsdf = NULL,
  prednames = NULL,
  SApackage = "JoSAE",
  SAmethod = "area",
  esttype = "TREE",
  totals = FALSE,
  estseed = "none",
  largebnd.unique = NULL,
  landarea = "FOREST",
  pcfilter = NULL,
  estvar = NULL,
  estvar.filter = NULL,
  allin1 = FALSE,
  metric = FALSE,
  variable.select = TRUE,
  estround = 3,
  pseround = 3,
  estnull = 0,
  psenull = "--",
  divideby = NULL,
  savedata = FALSE,
  savesteps = FALSE,
  rawdata = FALSE,
  rawonly = FALSE,
  multest = TRUE,
  addSAdomsdf = TRUE,
  SAdomvars = NULL,
  outfolder = NULL,
  outfn.pre = NULL,
  outfn.date = FALSE,
  addtitle = TRUE,
  raw_fmt = "csv",
  raw_dsn = "rawdata",
  savemultest = FALSE,
  multest_fmt = "csv",
  multest_outfolder = NULL,
  multest_dsn = NULL,
  multest_layer = NULL,
  multest.append = FALSE,
  multest.AOIonly = FALSE,
  overwrite_dsn = FALSE,
  overwrite_layer = TRUE,
  append_layer = FALSE,
  returntitle = FALSE,
  title.main = NULL,
  title.ref = NULL,
  title.dunitvar = NULL,
  title.estvar = NULL,
  title.filter = NULL,
  save4testing = FALSE,
  ...
)

Arguments

SApopdatlst

List. List of population data objects returned from modSApop().

SAdomsdf

DF/DT or comma-delimited file(*.csv). Dataframe from SAdoms with attributes from smallbnd to add to estimation table.

prednames

String vector. Name(s) of predictor variables to use in model.

SApackage

String. Small area package to use ('JoSAE', 'sae')

SAmethod

String. Small area method to use ('unit', 'area')

esttype

String. Estimation type ('TREE', 'AREA').

totals

Logical. If TRUE, returns total estimate (mean * AREAUSED).

estseed

String. Use seedling data only or add to tree data. Seedling estimates are only for counts (estvar='TPA_UNADJ')-('none', 'only', 'add').

largebnd.unique

String. Name of the large boundary unique identifer to define plots within a model extent. If NULL, all plots are used for model extent.

landarea

String. The sample area filter for estimates ('ALL', 'FOREST', 'TIMBERLAND'). If landarea=FOREST, filtered to COND_STATUS_CD = 1; If landarea=TIMBERLAND, filtered to SITECLCD in(1:6) and RESERVCD = 0.

pcfilter

String. A filter for plot or cond attributes (including pltassgn). Must be R logical syntax.

estvar

String. Name of the tree estimate variable.

estvar.filter

String. A tree filter for estimate variable. Must be R syntax (e.g., "STATUSCD == 1").

allin1

Logical. If TRUE, both estimates and percent sample error are output in one table as: estimates (percent sample error).

metric

Logical. If TRUE, output area is in metric units (hectares).

variable.select

Logical. If TRUE, selects useful predictors using mase:ElasticNet.

estround

Integer. Number of decimal places for estimates.

pseround

Integer. Number of decimal places for percent sampling error.

estnull

Number or character. The number or symbol to use to indicate 'not sampled' for estimate.

psenull

Number or character. The number or symbol to use to indicate 'not sampled' for percent standard error.

divideby

String. Conversion number for output ('hundred', 'thousand', 'million').

savedata

Logical. If TRUE, saves table(s) to outfolder.

savesteps

Logical. Saves graphs of predictors and response with labels whether selected or not for both area- and unit-level models.

rawdata

Logical. If TRUE, returns a list of raw data tables that are used for estimation (See Value). If savedata = TRUE, also written to outfolder.

rawonly

Logical. If TRUE, output rawdata only. If dataset includes many estimation units, and only raw data tables are needed, it is more efficient to output raw data only.

multest

Logical. If TRUE, returns a data frame of SA estimates using both unit-level and area-level estimates.

addSAdomsdf

Logical. If TRUE, sppends SAdomdf to unit.multest table for output.

SAdomvars

String vector. List of attributes from SAdoms to include in multest output.

outfolder

String. The outfolder to write files to. If NULL, files are written to working directory, or if gui, a window to browse.

outfn.pre

String. A prefix for outfile name, if savedata=TRUE.

outfn.date

Logical. If TRUE, add date to end of outfile (e.g., outfn_'date'.csv).

addtitle

Logical. If TRUE and savedata=TRUE, adds title to outfile.

raw_fmt

String. Format for raw output tables ('csv', 'sqlite', 'gpkg').

raw_dsn

String. Name of database if raw_fmt = c('sqlite', 'gpkg').

savemultest

Logical. If TRUE, save table with area- and unit-level estimates.

multest_fmt

String. Format for multest output tables ('csv', 'sqlite', 'gpkg').

multest_outfolder

String. Outfolder for multest. If NULL, same as outfolder.

multest_dsn

String. Name of database if multest_fmt = c('sqlite', 'gpkg').

multest_layer

String. Name of database layer if multest_fmt = c('sqlite', 'gpkg').

multest.append

Logical. If TRUE, appends multest dataframe to output.

multest.AOIonly

Logical. If TRUE, appends multest dataframe (AOI=1) to output.

overwrite_dsn

Logical. If TRUE, overwrites raw_dsn, if exists.

overwrite_layer

Logical. If TRUE, overwrites the output. If rawdata=TRUE, overwrites out_layer in rawdata folder (if raw_fmt = 'csv') or out_layers in raw_dsn (if raw_fmt != 'csv').

append_layer

Logical. If TRUE, and rawdata=TRUE, appends raw data to existing *.csv files (if raw_fmt = 'csv') or raw_dsn layers (if raw_fmt != 'csv".

returntitle

Logical. If TRUE, returns title(s) of the estimation table(s).

title.main

String. TITLE, if savedata=TRUE and/or returntitle=TRUE: the complete title used for table. If title.main=NULL, the title.* parameters are used to generate title string. Note: if title.ref is not NULL, it is added to title.main.

title.ref

String. TITLE, if savedata=TRUE and/or returntitle=TRUE: the ending text of the table title (e.g., Nevada, 2004-2005). If NULL, = "".

title.dunitvar

String. TITLE, if savedata=TRUE and/or returntitle=TRUE: pretty name for the estimation unit variable. If NULL, = unitvar.

title.estvar

String. TITLE: if savedata=TRUE and/or returntitle=TRUE: pretty name for the estimate variable. If NULL, title.estvar = estvar.name.

title.filter

String. TITLE, if savedata=TRUE and/or returntitle=TRUE: pretty name for filter(s). If title.filter=NULL, a default is generated from cfilter. If title.filter="", no title.filter is used.

save4testing

Logical. If TRUE, saves intermediate steps as R objects to outfolder for testing (pdomdat, dunitlut).

...

Parameters for modSApop() if SApopdat is NULL.

Value

est

Data frame. Tree estimates and percent sampling error by domain. Estimates are based on the SApackage and SAmethod parameters defined.

titlelst

List. List of titles used for table output.

raw

List of raw data. If rawdata=TRUE, a list including raw data components used for calculating estimate.

dunit.multest

Data frame. Table comparing different estimation strategies for SAE.

Raw data

domdat

Data frame. Domain-level data used for estimation.

estvar

String. Name of estimation variable.

estvar.filter

String. Logical filter specified for tree data.

dunit.totest

String. Table of estimates, including more details.

Details

If variables are NULL, then it will prompt user to input variables.

Necessary variables:

Data	Variable	Description

tree	tuniqueid	Unique identifier for each plot, to link to pltstrat (ex. PLT_CN).

CONDID	Unique identifier of each condition on plot, to link to cond. Set CONDID=1, if only 1 condition per plot.

TPA_UNADJ	Number of trees per acre each sample tree represents (ex. DESIGNCD=1: TPA_UNADJ=6.018046 for trees on subplot; 74.965282 for trees on microplot).

cond	cuniqueid	Unique identifier for each plot, to link to pltstrat (ex. PLT_CN).

CONDID	Unique identifier of each condition on plot. Set CONDID=1, if only 1 condition per plot.

CONDPROP_UNADJ	Unadjusted proportion of condition on each plot. Set CONDPROP_UNADJ=1, if only 1 condition per plot.

COND_STATUS_CD	Status of each forested condition on plot (i.e. accessible forest, nonforest, water, etc.)

NF_COND_STATUS_CD	If ACI=TRUE. Status of each nonforest condition on plot (i.e. accessible nonforest, nonsampled nonforest)

SITECLCD	If landarea=TIMBERLAND. Measure of site productivity.

RESERVCD	If landarea=TIMBERLAND. Reserved status.

SUBPROP_UNADJ	Unadjusted proportion of subplot conditions on each plot. Set SUBPROP_UNADJ=1, if only 1 condition per subplot.

MICRPROP_UNADJ	If microplot tree attributes. Unadjusted proportion of microplot conditions on each plot. Set MICRPROP_UNADJ=1, if only 1 condition per microplot.

MACRPROP_UNADJ	If macroplot tree attributes. Unadjusted proportion of macroplot conditions on each plot. Set MACRPROP_UNADJ=1, if only 1 condition per macroplot.

pltassign	puniqueid	Unique identifier for each plot, to link to cond (ex. CN).

STATECD	Identifies state each plot is located in.

INVYR	Identifies inventory year of each plot.

PLOT_STATUS_CD	Status of each plot (i.e. sampled, nonsampled). If not included, all plots are assumed as sampled.

Reference names are available for the following variables:
ADFORCD, AGENTCD, CCLCD, DECAYCD, DSTRBCD, KINDCD, OWNCD, OWNGRPCD, FORTYPCD, FLDTYPCD, FORTYPCDCALC, TYPGRPCD, FORINDCD, RESERVCD, LANDCLCD, STDSZCD, FLDSZCD, PHYSCLCD, MIST_CL_CD, PLOT_STATUS_CD, STATECD, TREECLCD, TRTCD, SPCD, SPGRPCD

Note

ADJUSTMENT FACTOR:
The adjustment factor is necessary to account for nonsampled conditions. For model-based estimation, we calculate adjustment factors by plot.

It is calculated by dividing 1 / summed condition proportions by plot. An adjustment factor is determined for each tree based on the size of the plot it was measured on. This is identified using TPA_UNADJ as follows:

PLOT SIZE	TPA_UNADJ

SUBPLOT	6.018046

MICROPLOT	74.965282

MACROPLOT	0.999188

If ACI=FALSE, only nonsampled forest conditions are accounted for in the adjustment factor.
If ACI=TRUE, the nonsampled nonforest conditions are removed as well and accounted for in adjustment factor. This is if you are interested in estimates for all lands or nonforest lands in the All-Condition-Inventory.

Common tree filters for estvar.filter:

FILTER	DESCRIPTION

"STATUSCD == 1"	Live trees

"STATUSCD == 2"	Dead trees

"TPAMORT_UNADJ > 0"	Mortality trees

"STATUSCD == 2 & DIA >= 5.0"	Dead trees >= 5.0 inches diameter

"STATUSCD == 2 & AGENTCD == 30"	Dead trees from fire

References

Breidenbach, J. 2018. JoSAE: Unit-Level and Area-Level Small Area Estimation. R package version 0.3.0. https://CRAN.R-project.org/package=JoSAE.

Molina I, Marhuenda Y. 2015. sae: An R Package for Small Area Estimation. The R Journal 7(1), 81-98. https://journal.r-project.org/archive/2015/RJ-2015-007/RJ-2015-007.

Author

Tracey S. Frescino, Paul L. Patterson, Elizabeth A. Freeman