Population data filters.

Returns a list of user-supplied parameters and parameter values for population data filters.

popFilters(
  evalid = NULL,
  states = NULL,
  evalCur = FALSE,
  evalEndyr = NULL,
  measCur = FALSE,
  measEndyr = NULL,
  invyrs = NULL,
  measyrs = NULL,
  intensity = NULL,
  ACI = FALSE,
  AOIonly = FALSE,
  ...
)

Arguments

evalid

Numeric. FIA Evaluation identifier for subsetting plots for population.

states

String or numeric vector. Name (e.g., 'Arizona','New Mexico') or code (e.g., 4, 35) of state(s) for evalid. If all states in one or more FIA Research Station is desired, set states=NULL and use RS argument to define RS.

evalCur

Logical. If TRUE, the most current FIA Evaluation is extracted for state(s).

evalEndyr

Number. The end year of the FIA Evaluation of interest. Selects only sampled plots and conditions for the evaluation period. If more than one state, create a named list object with evalEndyr labeled for each state (e.g., list(Utah=2014, Colorado=2013).

measCur

Logical. If TRUE, the most current sampled plots available for state(s).

measEndyr

Number. The most current sampled plots measured before or during end given..

invyrs

Integer vector. Inventory year(s) (e.g., c(2000, 2001, 2002)).

measyrs

Integer vector. Measurement year(s) (e.g., c(2000, 2001, 2002)).

intensity

Integer code. Code(s) indicating intensity to use for population.

ACI

Logical. If TRUE, including All Condition Inventory (ACI) plots.

AOIonly

Logical. If TRUE, and there is an AOI (1/0) attribute in the population data, only AOI=1 are used for estimation.

...

For extendibility.

Value

A list of user-supplied parameters and parameter values for population data filters.

Details

If no parameters, an empty list is returned.

Author

Grayson W. White

Examples

popFilters(ACI = TRUE)
#> $evalid
#> NULL
#> 
#> $states
#> NULL
#> 
#> $evalCur
#> [1] FALSE
#> 
#> $evalEndyr
#> NULL
#> 
#> $measCur
#> [1] FALSE
#> 
#> $measEndyr
#> NULL
#> 
#> $invyrs
#> NULL
#> 
#> $measyrs
#> NULL
#> 
#> $intensity
#> NULL
#> 
#> $ACI
#> [1] TRUE
#> 
#> $AOIonly
#> [1] FALSE
#>