Spatial - Aligns a list of raster layer(s) based on a reference raster.

Rasters are pixel-aligned and reprojected using the gdal warp function with help from the gdalraster package. The extent of the reference raster is used or a given boundary extent.

Usage

spAlignRast(
  ref_rastfn,
  rastlst,
  resample_methodlst = NULL,
  clip = FALSE,
  bnd = NULL,
  bnd_dsn = NULL,
  tile = TRUE,
  tile_blocksize = 256,
  makestack = FALSE,
  outrastnmlst = NULL,
  outfolder = NULL,
  overwrite = TRUE
)

Arguments

ref_rastfn

String. Full path name of reference raster.

rastlst

String. Full path names of one or more rasters to align.

resample_methodlst

String. Resample method ('mode', 'near', 'bilinear', 'cubic', 'cubicspline', 'max', 'min', 'med', 'average'). Suggested values: if raster type is categorical; 'mode' or 'near'. if raster type is continuous; 'bilinear', 'cubic'.

clip

Logical. If TRUE, subset raster to a boundary.

bnd

R object or Full path name to a shapefile or layer in a database.

bnd_dsn

String. Data source name of bnd, if bnd is a layer in a database.

tile

Logical. If TRUE, tile the output raster.

tile_blocksize

Numeric. If tile = TRUE, define the size of tile block.

makestack

Logical. If TRUE, makes a raster stack with format 'GTIFF'.

outrastnmlst

String. Base name of output raster (e.g., 'elev').

outfolder

String. Name of folder for writing output raster. If NULL, outfolder = getwd().

overwrite

Logical. If TRUE, overwrite output raster.

Value

String. List of output raster file names.