Spatial wrapper - Extracts XY coordinates within a given boundary.

Wrapper to get FIA plots within the boundary population (area of interest) - Intersect with state boundary - Get FIA plots for intersected states, including tree, and spatial - Clip spatial coordinates and other tables to boundary (spClipPoint)

spGetXY(
  bnd,
  bnd_dsn = NULL,
  bnd.filter = NULL,
  states = NULL,
  RS = NULL,
  xy_datsource = "datamart",
  xy = NULL,
  xy_dsn = NULL,
  xy.uniqueid = "PLT_CN",
  xvar = NULL,
  yvar = NULL,
  xy.crs = 4269,
  xyjoinid = NULL,
  pjoinid = "CN",
  clipxy = TRUE,
  plot_layer = "plot",
  evalid = NULL,
  evalCur = FALSE,
  evalEndyr = NULL,
  measCur = FALSE,
  measEndyr = NULL,
  measEndyr.filter = NULL,
  invyrs = NULL,
  measyrs = NULL,
  allyrs = FALSE,
  intensity1 = FALSE,
  showsteps = FALSE,
  returnxy = TRUE,
  savedata = FALSE,
  exportsp = FALSE,
  savedata_opts = NULL
)

Arguments

bnd

sf R object, Area of Interest (AOI) boundary. Can be a spatial sf object, full pathname to a shapefile, or name of a layer within a database.

bnd_dsn

String. Data source name (dsn; e.g., SQLite database or shapefile pathname) of bnd. The dsn varies by driver. See gdal OGR vector formats (https://www.gdal.org/ogr_formats.html). Optional if bnd is an R object.

bnd.filter

String. Filter to subset bnd spatial layer.

states

String. The name of state(s) for tables (e.g., "Vermont", "Utah").

RS

String. Name of FIA research station to restrict states to ('RMRS','SRS','NCRS','NERS','PNWRS'). If NULL, all research stations are included.

xy_datsource

String. Source of XY data ("obj", "csv", "datamart", "sqlite"). If datsource=NULL, checks extension of xy_dsn or xy to identify datsource.

xy

sf R object or String. Table with xy coordinates. Can be a spatial polygon object, data frame, full pathname to a shapefile, or name of a layer within a database.

xy_dsn

String. Data source name (dsn; i.e., pathname or database name) of xy. The dsn varies by driver. See gdal OGR vector formats (https://www.gdal.org/ogr_formats.html). Optional if bnd_layer is an R object.

xy.uniqueid

String. Unique identifier of xy.

xvar

String. Name of variable in xyplt defining x coordinate.

yvar

String. Name of variable in xyplt defining y coordinate.

xy.crs

PROJ.4 String or CRS object or Integer EPSG code defining Coordinate Reference System.

xyjoinid

String. Variable in xy to join to plot data. If NULL, xyjoinid=xy.uniqueid.

pjoinid

String. Variable in plt to join to XY data. Not necessary to be unique. If using most current XY coordinates, use identifier for a plot (e.g., PLOT_ID).

clipxy

Logical. If TRUE, clips xy data to bnd.

plot_layer

String. Name of plot_layer in database.

evalid

Integer. To extract data for a specific evaluation period. See notes for more information about FIA Evaluations.

evalCur

Logical. If TRUE, extract plots with most current FIA Evalidation for state(s).

evalEndyr

Integer. Defining end year of Evaluation (yyyy).

measCur

Logical. If TRUE, extract plots with most current measurement for state(s).

measEndyr

Integer year (YYYY). If measCur=TRUE, extract plots with most current measurement for state(s) for years measured before measEndyr.

measEndyr.filter

Filter. If measCur=TRUE and measEndyr != NULL, a filter for bnd to identify and area to use measEndyr, such as disturbed areas where you want to exclude plots measured after disturbance.

invyrs

Integer vector. Defining specific inventory years of data (e.g., 2010:2015).

measyrs

Integer vector. Defining specific measurement years of data (e.g., 2010:2015).

allyrs

Logical. If TRUE, selects all years (annual inventory) in database.

intensity1

Logical. If TRUE, include only single intensity plots (i.e., INTENSITY = 1).

showsteps

Logical. If TRUE, display data in device window.

returnxy

Logical. If TRUE, returns XY coordinates.

savedata

Logical. If TRUE, saves data to outfolder. Note: includes XY data if returnxy = TRUE.

exportsp

Logical. If savedata = TRUE and returnxy = TRUE, if TRUE, exports XY data as spatial data.

savedata_opts

List. See help(savedata_options()) for a list of options. Only used when savedata = TRUE.

Value

spxy

sf. If returnxy=TRUE, spatial xy point data.

pltids

data frame. A table of pltids that are within bnd.

spxy

sf data frame. If returnxy, a simple feature with pltids within bnd.

bndx

sf object. Input bnd.

xy.uniqueid

String. Unique identifier of plots in xy.

states

String. Vector of states that intersect bnd.

countyfips

String. Vector of countyfips values that intersect bnd.

stbnd.att

String. Name of state attribute used to select plots.

If savedata=TRUE and returnxy=TRUE, the plt data frame, including XY coordinates is saved to outfolder (xyplt).
If savedata=TRUE and returnxy=FALSE, the plt data frame, without XY coordinates is saved to outfolder (pltids).
If savedata=TRUE and returnxy=TRUE and exportsp=TRUE, the spxy sf object is exported as shapefile to outfolder.

Details

datsource

Plots are extracted from 3 different data sources:
1) CSV - data have previously been extracted from the FIA database and stored as CSV files.
2) datamart - data are extracted from FIA's publically-available datamart.
3) sqlite - data have previously been extracted from the FIA database and stored within a SQLite database.

Selection parameters

FIA plots are selected based on the following parameters:

evalid - the FIA evaluation identifier

evalCur - the most current FIA evaluation in database

evalEndyr - the FIA evaluation ending in evalEndyr

evalType - the FIA evaluation type ('ALL', 'AREAVOL', 'GRM', 'P2VEG', 'DWM', 'INV', 'REGEN', 'CRWN')

measCur - the most current measurement of each plot in database

measEndyr - the most current measuremtn of each plot in database in or prior to measEndyr

Endyr.filter - a filter for bnd that specifies the boundary where measEndyr should be applied

Note

If savebnd=TRUE:
If out_fmt=c('csv','shp'), the writeOGR (rgdal) function is called. The ArcGIS driver truncates variable names to 10 characters or less. Variable names are changed before export using an internal function (trunc10shp). If Spatial object has more than 1 record, it will be returned but not exported.

If datsource="datmart", (default), data are imported from FIA DataMart. The plot coordinates have been altered for privacy (See https://www.fia.fs.fed.us/tools-data/spatial/Policy/index.php for details). The zip files are extracted on-the-fly from the online website. Web server connections will affect download speeds.

Author

Tracey S. Frescino

Examples

if (FALSE) {
# Set up data from FIESTA
WYbhfn <- system.file("extdata",
                      "sp_data/WYbighorn_adminbnd.shp",
                      package = "FIESTA")
# Use spGetXY
WYbhxy <- spGetXY(bnd = WYbhfn,
                  xy_datsource = "datamart",
                  evalCur = TRUE,
                  returnxy = TRUE)
}