Unit options.

Returns a list of user-supplied parameters and parameter values for unit.

unit_options(
  unitvar2 = NULL,
  areaunits = "acres",
  minplotnum.unit = 10,
  unit.action = "keep",
  npixelvar = "npixels",
  ...
)

Arguments

unitvar2

String. Name of a second level estimation unit variable in unitarea and cond or pltassgn with assignment for each plot (e.g., 'STATECD').

areaunits

String. Units of areavar in unitarea ('acres', 'hectares').

minplotnum.unit

Integer. Minimum number of plots for estimation unit.

unit.action

String. What to do if number of plots in an estimation unit is less than minplotnum.unit ('keep', 'remove' 'combine'). If unit.action='combine', combines estimation unit to the following estimation unit, ordered in unitzonal or stratalut.

npixelvar

String. Name of variable in unitlut defining number of pixels by estimation unit.

...

For extendibility.

Value

A list of user-supplied parameters and parameter values for strata.

Details

If no parameters, an empty list is returned.

Author

Grayson W. White

Examples

unit_options()
#> $formallst
#> [1] "unitvar2"        "areaunits"       "minplotnum.unit" "unit.action"    
#> [5] "npixelvar"       "..."            
#> 
#> $input.params
#> NULL
#> 
#> $unitvar2
#> NULL
#> 
#> $areaunits
#> [1] "acres"
#> 
#> $minplotnum.unit
#> [1] 10
#> 
#> $unit.action
#> [1] "keep"
#> 
#> $npixelvar
#> [1] "npixels"
#>