Data - Gets variable description or class for SPCD.

Merge the ref_species table to append new variables, names, or categories to x.

datLUTspp(
  x,
  xvar = "SPCD",
  NAclass = "Other",
  group = FALSE,
  states = NULL,
  name = "COMMON",
  add0 = FALSE,
  stopifmiss = FALSE,
  xtxt = NULL,
  savedata = FALSE,
  savedata_opts = NULL,
  gui = FALSE
)

Arguments

x

Data frame or comma-delimited file (*.csv). The data table with variable to classify.

xvar

String. Name of variable in the data table to join to.

NAclass

String. NA values in xvar will be changed to NAclass.

group

Logical. If TRUE, the group variable in ref_species are merged to data table (E_SPGRPCD, W_SPGRPCD), depending on state(s) specified. If states overlap both E and W regions, the region with majority is used or E if equal. The group name is merged from ref_codes, SPGRPCD Variable.

states

String. Name of state(s) the x table is from.

name

String. Name for species output type ('COMMON', 'SCIENTIFIC', 'SYMBOL').

add0

Logical. IF TRUE, keep all codes in look up table. If FALSE, only include codes that are in x.

stopifmiss

Logical. IF TRUE, stops function if missing codes in LUTx.

xtxt

String.* Name of x table for more useful information in warnings.

savedata

Logical. If TRUE, saves data to outfolder.

savedata_opts

List. See help(savedata_options()) for a list of options. Only used when savedata = TRUE. If out_layer = NULL, default = 'datlut'.

gui

Logical. If gui, user is prompted for parameters.

Value

xLUT

The input data table with look-up table variable(s).

xLUTnm

Name of the new variable(s).

LUT

Look up table with categories.

If savedata = TRUE, a comma-delimited file is output to the outfolder as outfn. If outfn = NULL, the name of the file will be datlut_'date'.csv.

Note

For available reference tables: sort(unique(ref_codes$VARIABLE))

Author

Tracey S. Frescino

Examples

WYtreelut <- datLUTspp(WYtree)
names(WYtreelut)
#> [1] "xLUT"    "xLUTnm"  "LUT"     "ref_spp"
WYtree2 <- WYtreelut$xLUT
head(WYtree2)
#>    SPCD         PLT_CN CONDID SUBP TREE STATUSCD SPGRPCD DIA HT TREECLCD
#> 1:   19 40404738010690      1    2    3        2      12  NA NA       NA
#> 2:   19 40404738010690      1    2    4        2      12  NA NA       NA
#> 3:   19 40404738010690      1    2    7        1      12 5.3 26        2
#> 4:   19 40404821010690      1    1    7        1      12 5.9 48        2
#> 5:   19 40404821010690      1    1    8        1      12 8.1 45        2
#> 6:   19 40404821010690      1    1   11        2      12  NA NA       NA
#>    AGENTCD STANDING_DEAD_CD VOLCFNET VOLCFGRS VOLBFNET TPA_UNADJ DRYBIO_AG
#> 1:      70                0       NA       NA       NA        NA   0.00000
#> 2:      70                0       NA       NA       NA        NA   0.00000
#> 3:      NA               NA 0.864289 0.864289       NA  6.018046  25.48605
#> 4:      NA               NA 3.046342 3.046342       NA  6.018046  88.91319
#> 5:      NA               NA 6.019584 6.019584       NA  6.018046 171.35480
#> 6:      70                0       NA       NA       NA        NA   0.00000
#>    CARBON_AG        BA   COMMON_NAME
#> 1:        NA        NA subalpine fir
#> 2:        NA        NA subalpine fir
#> 3:  12.74303 0.1532029 subalpine fir
#> 4:  44.45660 0.1898537 subalpine fir
#> 5:  85.67740 0.3578369 subalpine fir
#> 6:        NA        NA subalpine fir