Data - Generates a pivot table.

Generates a pivot table of values by x row and y column.

datPivot(
  x,
  pvar,
  xvar,
  yvar,
  pfun = sum,
  xfilter = NULL,
  NAto0 = TRUE,
  dropNAxvar = TRUE,
  dropNAyvar = TRUE,
  pvar.round = 2,
  savedata = FALSE,
  savedata_opts = NULL,
  gui = FALSE
)

Arguments

x

Dataframe. Table with pivot variables.

pvar

String. The name of the variable for pivot table values.

xvar

String. The name of the variable for rows.

yvar

String. The name of the variable for columns.

pfun

Function. The name of the function to use for pivot table values (ex. sum, mean, max).

xfilter

String. A filter to subset the datatable table x before pivoting (ex. "STATUSCD == 1").

NAto0

Logical. If TRUE, converts NA values to 0.

dropNAxvar

Logical. If TRUE, removes columns that are NA.

dropNAyvar

Logical. If TRUE, removes rows that have NA values.

pvar.round

Integer. Number to round pvar values to.

savedata

Logical. If TRUE, writes output data to outfolder.

savedata_opts

List. See help(savedata_options()) for a list of options. Only used when savedata = TRUE. If out_layer = NULL, default = 'datpivot'.

gui

Logical. If TRUE, pop-up windows will appear for user-interface.

Value

ptab

Matrix. The pivot table.

Author

Tracey S. Frescino

Examples

# Pivot WYcond table
datPivot(x = FIESTA::WYcond,
         pvar = "CONDPROP_UNADJ",
         xvar = "FORTYPCD",
         yvar = "STDSZCD")
#>     FORTYPCD     1     2     3     5
#>  1:      182 23.23  3.00  4.50  0.00
#>  2:      184 15.50  0.00  1.00  0.00
#>  3:      185  0.75  0.00  0.00  0.00
#>  4:      201 27.85  5.50  8.50  0.00
#>  5:      221 39.97  2.25  2.75  0.00
#>  6:      265 16.25  3.25  3.25  0.00
#>  7:      266 48.25 12.51 10.50  0.00
#>  8:      268 28.38  7.50  9.49  0.00
#>  9:      269  1.00  0.00  0.00  0.00
#> 10:      281 43.95 39.43 34.68  0.00
#> 11:      366  3.50  0.00  7.59  0.00
#> 12:      367  4.37  5.75  7.50  0.00
#> 13:      509  1.00  1.75  2.00  0.00
#> 14:      517  0.00  0.00  1.00  0.00
#> 15:      703  3.03  0.25  1.00  0.00
#> 16:      706  0.00  0.50  0.00  0.00
#> 17:      901  2.00  8.12 19.25  0.00
#> 18:      999  0.00  0.00  0.00 40.75

# Pivot WYtree table
datPivot(x = FIESTA::WYtree,
         pvar = "TPA_UNADJ",
         xvar = "SPCD",
         yvar = "STATUSCD",
         pfun = mean,
         NAto0 = TRUE)
#>     SPCD     1    2
#>  1:  101 17.17 6.02
#>  2:  106  6.02 0.00
#>  3:  108 15.72 6.02
#>  4:  113 17.74 6.02
#>  5:  122 12.10 6.02
#>  6:   19 20.60 6.02
#>  7:  202 11.54 6.02
#>  8:  313  6.02 6.02
#>  9:  375 74.97 6.02
#> 10:  475  6.02 6.02
#> 11:  544  6.02 0.00
#> 12:   65  8.01 6.02
#> 13:   66 13.21 6.02
#> 14:  745  6.02 6.02
#> 15:  746 22.97 6.02
#> 16:  749  8.67 6.02
#> 17:  823 13.94 6.02
#> 18:   93 14.15 6.02
#> 19:   96 29.00 0.00