Data - Aggregates numeric tree data to the plot or condition-level.

Aggregates numeric tree-level data (e.g., VOLCFNET) to plot or condition, including options for filtering tree data or extrapolating to plot acre by multiplying by TPA.

datSumTree(
  tree = NULL,
  seed = NULL,
  cond = NULL,
  datsource = "obj",
  data_dsn = NULL,
  plt = NULL,
  plt_dsn = NULL,
  tuniqueid = "PLT_CN",
  cuniqueid = "PLT_CN",
  puniqueid = "CN",
  bycond = FALSE,
  condid = "CONDID",
  bysubp = FALSE,
  subpid = "SUBP",
  tsumvarlst = NULL,
  tsumvarnmlst = NULL,
  TPA = TRUE,
  tfun = sum,
  ACI = FALSE,
  tfilter = NULL,
  addseed = FALSE,
  lbs2tons = TRUE,
  metric = FALSE,
  getadjplot = FALSE,
  adjtree = FALSE,
  adjvar = "tadjfac",
  adjTPA = 1,
  NAto0 = FALSE,
  tround = 5,
  checkNA = FALSE,
  returnDT = TRUE,
  savedata = FALSE,
  savedata_opts = NULL,
  gui = FALSE
)

Arguments

tree

Dataframe or comma-delimited file (*.csv). The tree-level table.

seed

Dataframe or comma-delimited file (*.csv). The seedling table.

cond

Dataframe or comma-delimited file (*.csv). Condition-level table to join the aggregated tree data to, if bycond=TRUE. This table also may be used for condition proportion or strata variables used if adjcond or adjstrata = TRUE (See details below). This table is optional.

datsource

String. Source of data ('obj', 'csv', 'sqlite', 'gdb').

data_dsn

String. If datsource='sqlite', the name of SQLite database (*.sqlite).

plt

Dataframe, comma-delimited file (*.csv), or shapefile (*.shp). Plot-level table to join the aggregated tree data to, if bycond=FALSE. This table is optional.

plt_dsn

String. The data source name (dsn; i.e., folder or database name) of plt. The dsn varies by driver. See gdal OGR vector formats (https://www.gdal.org/ogr_formats.html). Optional.

tuniqueid

String. Unique identifier of plt in tree table.

cuniqueid

String. Unique identifier of plt in cond table if cond is NOT NULL.

puniqueid

String. Unique identifier of plt table if plt is NOT NULL.

bycond

Logical. If TRUE, the data are aggregated to the condition level (by: cuniqueid, condid). If FALSE, the data are aggregated to the plot level (by: puniqueid).

condid

String. Unique identifier for conditions.

bysubp

Logical. If TRUE, data are aggregated to the subplot level.

subpid

String. Unique identifier of each subplot.

tsumvarlst

String (vector). Tree-level variable(s) to aggregate (e.g., "TPA_UNADJ", "BA"). Use tsumvar="PLT_CN" (tfun=sum) for summed tree count.

tsumvarnmlst

String (vector). Name of the tree-level variable(s) to aggregate (e.g., "TPALIVE", "BALIVE"). This list must have the same number of variables as tsumvarlst and be in respective order. If NULL, the default names will be tsumvar'_tfun' (e.g., "TPA_UNADJ_SUM", "BA_SUM").

TPA

Logical. If TRUE, tsumvarlst variable(s) are multiplied by the respective trees-per-acre variable (see details) to get per-acre measurements.

tfun

Function. The name of the function to use to aggregate the data (e.g., sum, mean, max).

ACI

Logical. If TRUE, if ACI (All Condition Inventory) plots exist, any trees on these plots will be included in summary. If FALSE, you must include condition table.

tfilter

String. Filter to subset the tree data before aggregating (e.g., "STATUSCD == 1"). This must be in R syntax. If tfilter=NULL, user is prompted. Use tfilter="NONE" if no filters.

addseed

Logical. If TRUE, add seedling data to tree counts (if TPA variable in tsumvarlst).

lbs2tons

Logical. If TRUE, converts biomass or carbon variables from pounds to tons.

metric

Logical. If TRUE, converts response to metric units based on FIESTA::ref_conversion, if any variable in tsumvarlst is in FIESTAutils::ref_estvar. Note: if TPA, TPA is converted to trees per hectare (TPH: 1 / (1/ tpavar * 0.4046860)).

getadjplot

Logical. If TRUE, adjustments are calculated for nonsampled conditions on plot.

adjtree

Logical. If TRUE, trees are individually adjusted by adjustment factors. Adjustment factors must be included in tree table (see adjvar).

adjvar

String. If adjtree=TRUE, the name of the variable to use for multiplying by adjustment (e.g., tadjfac).

adjTPA

Numeric. A tree-per-acre adjustment. Use for DESIGNCD=1 (annual inventory), if using less than 4 subplots. If using only 1 sublot for estimate, adjTPA=4. The default is 1.

NAto0

Logical. If TRUE, convert NA values to 0.

tround

Number. The number of digits to round to. If NULL, default=6.

checkNA

Logical. If TRUE, checks if NA values exist in necessary variables.

returnDT

Logical. If TRUE, returns data.table object(s). If FALSE, returns data.frame object(s).

savedata

Logical. If TRUE, saves data to outfolder.

savedata_opts

List. See help(savedata_options()) for a list of options. Only used when savedata = TRUE. If out_layer = NULL, default = 'treesum'.

gui

Logical. If gui, user is prompted for parameters.

Value

A list of the following items:

treedat

Data frame. Plot or condition-level table with aggregated tree attributes.

sumvars

String vector. Name(s) of the output aggregated tree attributes.

If savedata=TRUE
- treedat will be saved to the outfolder.
- a text file of input parameters is saved to outfolder ('outfn'_parameters_'date'.txt).

Details

If variable = NULL, then it will prompt user for input.

Dependent external functions: datFilter Dependent internal functions: addcommas, fileexistsnm, getadjfactor

For adjcond (bycond=FALSE):
If you want to get trees-per-acre information aggregated to plot or condition level, you need to include a TPA variable in tree table.
For tsumvars = GROWCFGS, GROWBFSL, GROWCFAL, FGROWCFGS, FGROWBFSL, or FGROWCFAL, you must have TPAGROW_UNADJ
For tsumvars = MORTCFGS, MORTBFSL, MORTCFAL, FMORTCFGS, FMORTBFSL, or FMORTCFAL, you must have TPAMORT_UNADJ
For tsumvars = REMVCFGS, REMVBFSL, REMVCFAL, FREMVCFGS, FREMVBFSL, or FREMVCFAL, you must have TPAREMV_UNADJ

If you want to adjust plot-level information by condition proportions (adjplot), you need to include CONDID & CONDPROP_UNADJ in cond or tree table and COND_STATUS_CD and FORTYPCD if adjplot="FVS". For adjplot="WEIGHTED" - each tree is weighted (or multiplied) by the condition proportions (CONDPROP_UNADJ). For adjplot="FVS", each tree is weighted (or multiplied) by the sum of the condition proportions (CONDPROP_UNADJ) that are forested and not nonstocked (COND_STATUS_CD = 1 and FORTYPCD != 999).

Note

If a dat table is provided, the aggregated tree data will be merged to table and NULL values will be output as 0.

Author

Tracey S. Frescino

Examples

# Aggregate LIVE_CANOPY_CVR_PCT to plot, weighted by CONDPROP_UNADJ
treesum <- datSumTree(tree = FIESTA::WYtree,
                      tsumvarlst = "TPA_UNADJ")$treedat
#> Warning: COND_STATUS_CD not in table, assuming forested plots with no ACI plots

# Check results
treesum[treesum$PLT_CN == 40404737010690,]
#>            PLT_CN COUNT_TPA
#> 1: 40404737010690  12.03609
FIESTA::WYtree[FIESTA::WYtree$PLT_CN == 40404737010690,]
#>               PLT_CN CONDID SUBP TREE STATUSCD SPCD SPGRPCD DIA HT TREECLCD
#> 17431 40404737010690      2    3    1        2  746      44  NA NA       NA
#> 17432 40404737010690      2    3    2        2  746      44  NA NA       NA
#> 17433 40404737010690      2    3    3        1  746      44 5.9 24        2
#> 17434 40404737010690      2    3    6        1  746      44 6.3 22        3
#> 17435 40404737010690      2    3    7        2  746      44  NA NA       NA
#> 17436 40404737010690      2    3    8        2  746      44  NA NA       NA
#>       AGENTCD STANDING_DEAD_CD VOLCFNET VOLCFGRS VOLBFNET TPA_UNADJ DRYBIO_AG
#> 17431      70                0       NA       NA       NA        NA   0.00000
#> 17432      70                0       NA       NA       NA        NA   0.00000
#> 17433      NA               NA 0.852096 0.926191       NA  6.018046  34.61910
#> 17434      NA               NA 0.807224 1.009030       NA  6.018046  32.16685
#> 17435      70                0       NA       NA       NA        NA   0.00000
#> 17436      70                0       NA       NA       NA        NA   0.00000
#>       CARBON_AG        BA
#> 17431        NA        NA
#> 17432        NA        NA
#> 17433  17.30955 0.1898537
#> 17434  16.08343 0.2164693
#> 17435        NA        NA
#> 17436        NA        NA