Population data filters.

Returns a list of user-supplied parameters and parameter values for population data filters.

popFilters(
  evalid = NULL,
  invyrs = NULL,
  intensity = NULL,
  measCur = FALSE,
  measEndyr = NULL,
  measEndyr.filter = NULL,
  ACI = FALSE,
  ...
)

Arguments

evalid

Numeric. FIA Evaluation identifier for subsetting plots for population.

invyrs

Integer vector. Inventory year(s) (e.g., c(2000, 2001, 2002)).

intensity

Integer code. Code(s) indicating intensity to use for population.

measCur

Logical. Inventory span defining variable. If TRUE, extract plots with most current measurement for state(s).

measEndyr

Logical. Inventory span defining variable. If TRUE, extract plots with most current measurement for state(s) for years measured in or before measEndyr.

measEndyr.filter

Filter for extracting plots using measEndyr. Must be in R syntax (e.g., 'AOI == 1').

ACI

Logical. If TRUE, including All Condition Inventory (ACI) plots.

...

For extendibility.

Value

A list of user-supplied parameters and parameter values for population data filters.

Details

If no parameters, an empty list is returned.

Author

Grayson W. White

Examples

popFilters(ACI = TRUE)
#> $evalid
#> NULL
#> 
#> $invyrs
#> NULL
#> 
#> $intensity
#> NULL
#> 
#> $measCur
#> [1] FALSE
#> 
#> $measEndyr
#> NULL
#> 
#> $measEndyr.filter
#> NULL
#> 
#> $ACI
#> [1] TRUE
#>