Spatial wrapper - Extracts plot data within a given boundary.

Wrapper to get FIA plots within the boundary population (area of interest) - Intersect with state boundary - Get FIA plots for intersected states, including tree, and spatial - Clip spatial coordinates and other tables to boundary (spClipPoint)

spGetPlots(
  bnd = NULL,
  bnd_dsn = NULL,
  bnd.filter = NULL,
  states = NULL,
  RS = NULL,
  pltids = NULL,
  xy_datsource = NULL,
  xy = NULL,
  xy_dsn = NULL,
  xy.uniqueid = "PLT_CN",
  xvar = NULL,
  yvar = NULL,
  xy.crs = 4269,
  xyjoinid = NULL,
  pjoinid = NULL,
  clipxy = TRUE,
  datsource = NULL,
  data_dsn = NULL,
  istree = FALSE,
  isseed = FALSE,
  isveg = FALSE,
  isdwm = FALSE,
  plot_layer = "plot",
  cond_layer = "cond",
  tree_layer = "tree",
  seed_layer = "seedling",
  vsubpspp_layer = "p2veg_subplot_spp",
  vsubpstr_layer = "p2veg_subp_structure",
  invsubp_layer = "invasive_subplot_spp",
  dwm_layer = "cond_dwm_calc",
  ppsa_layer = "pop_plot_stratum_assgn",
  other_layers = NULL,
  puniqueid = "CN",
  pltassgnid = "PLT_CN",
  savePOP = FALSE,
  evalid = NULL,
  evalCur = FALSE,
  evalEndyr = NULL,
  evalType = "VOL",
  measCur = FALSE,
  measEndyr = NULL,
  measEndyr.filter = NULL,
  invyrs = NULL,
  measyrs = NULL,
  allyrs = FALSE,
  intensity1 = FALSE,
  showsteps = FALSE,
  returndata = TRUE,
  savedata = FALSE,
  savebnd = FALSE,
  returnxy = TRUE,
  exportsp = FALSE,
  savedata_opts = NULL,
  spXYdat = NULL,
  gui = FALSE
)

Arguments

bnd

sf R object, Area of Interest (AOI) boundary. Can be a spatial sf object, full pathname to a shapefile, or name of a layer within a database.

bnd_dsn

String. Data source name (dsn; e.g., SQLite database or shapefile pathname) of bnd. The dsn varies by driver. See gdal OGR vector formats (https://www.gdal.org/ogr_formats.html). Optional if bnd is an R object.

bnd.filter

String. Filter to subset bnd spatial layer.

states

String. The name of state(s) for tables (e.g., "Vermont", "Utah").

RS

String. Name of FIA research station to restrict states to ('RMRS','SRS','NCRS','NERS','PNWRS'). If NULL, all research stations are included.

pltids

Data frame. Non-spatial plot identifiers within bnd).

xy_datsource

String. Source of XY data ("obj", "csv", "datamart", "sqlite"). If datsource=NULL, checks extension of xy_dsn or xy to identify datsource.

xy

sf R object or String. Table with xy coordinates. Can be a spatial polygon object, data frame, full pathname to a shapefile, or name of a layer within a database.

xy_dsn

String. Data source name (dsn; i.e., pathname or database name) of xy. The dsn varies by driver. See gdal OGR vector formats (https://www.gdal.org/ogr_formats.html). Optional if bnd_layer is an R object.

xy.uniqueid

String. Unique identifier of xy.

xvar

String. Name of variable in xyplt defining x coordinate.

yvar

String. Name of variable in xyplt defining y coordinate.

xy.crs

PROJ.4 String or CRS object or Integer EPSG code defining Coordinate Reference System.

xyjoinid

String. Variable in xy to join to plot data. If NULL, xyjoinid=xy.uniqueid.

pjoinid

String. Variable in plt to join to XY data. Not necessary to be unique. If using most current XY coordinates, use identifier for a plot (e.g., PLOT_ID).

clipxy

Logical. If TRUE, clips xy data to bnd.

datsource

String. Source of FIA data ("obj", "csv", "datamart", "sqlite"). If datsource="sqlite", specify database name in data_dsn and layers in *_layer arguments. If datsource="datamart", files are downloaded and extracted from FIA DataMart (http://apps.fs.usda.gov/fia/datamart/datamart.html). See details for more information about plot coordinates. If datsource="csv", specify *.csv file names in *_layer arguments.

data_dsn

String. Name of database where *_layers reside.

istree

Logical. If TRUE, extract tree data from FIA database.

isseed

Logical. If TRUE, extract seedling data from FIA database.

isveg

Logical. If TRUE, understory vegetation tables are extracted from FIA database (P2VEG_SUBPLOT_SPP, P2VEG_SUBP_STRUCTURE, INVASIVE_SUBPLOT_SPP).

isdwm

Logical. If TRUE, down woody material are extracted.

plot_layer

String. Name of layer in database of file name of FIA plot table.

cond_layer

String. Name of layer in database of file name of FIA cond table.

tree_layer

String. Name of layer in database of file name of FIA tree table if istree=TRUE.

seed_layer

String. Name of layer in database of file name of FIA seedling table if isseed=TRUE.

vsubpspp_layer

String. Name of layer in database of file name of FIA P2VEG_SUBPLOT_SPP table if isveg=TRUE.

vsubpstr_layer

String. Name of layer in database of file name of FIA P2VEG_SUBP_STRUCTURE table if isveg=TRUE.

invsubp_layer

String. Name of layer in database of file name of FIA INVASIVE_SUBPLOT_SPP table if isveg=TRUE.

dwm_layer

String. Name of layer in database of file name of FIA COND_DWM_CALC table if isdwm=TRUE.

ppsa_layer

String. Name of layer in database of file name of FIA pop_plot_stratum_assgn table, if using evaluations.

other_layers

String. Other layer(s) in database to clip and/or extract from database (Note: must include PLT_CN variable as unique identifier).

puniqueid

String. Name of unique identifier of plt.

pltassgnid

String. Name of unique identifier of pop_plot_stratum_assgn.

savePOP

Logical. If TRUE, returns and/or saves POP_PLOT_STRATUM_ASSGN table.

evalid

Integer. To extract data for a specific evaluation period. See notes for more information about FIA Evaluations.

evalCur

Logical. If TRUE, extract plots with most current FIA Evalidation for state(s).

evalEndyr

Integer. Defining end year of Evaluation (yyyy).

evalType

String vector. The type(s) of evaluation of interest ('ALL', 'AREAVOL', 'GRM', 'P2VEG', 'DWM', 'INV', 'REGEN', 'CRWN'). The evalType 'ALL' includes nonsampled plots; 'AREAVOL' includes plots used for area or tree estimates (eval_typ plots used for growth, removals, mortality, and change estimates (eval_typ database manual for regional availability and/or differences.

measCur

Logical. If TRUE, extract plots with most current measurement for state(s).

measEndyr

Integer year (YYYY). If measCur=TRUE, extract plots with most current measurement for state(s) for years measured before measEndyr.

measEndyr.filter

Filter. If measCur=TRUE and measEndyr != NULL, a filter for bnd to identify and area to use measEndyr, such as disturbed areas where you want to exclude plots measured after disturbance.

invyrs

Integer vector. Defining specific inventory years of data (e.g., 2010:2015).

measyrs

Integer vector. Defining specific measurement years of data (e.g., 2010:2015).

allyrs

Logical. If TRUE, selects all years (annual inventory) in database.

intensity1

Logical. If TRUE, include only single intensity plots (i.e., INTENSITY = 1).

showsteps

Logical. If TRUE, display data in device window.

returndata

Logical. If TRUE, returns data objects.

savedata

Logical. If TRUE, saves data to outfolder.

savebnd

Logical. If TRUE, saves bnd. If out_fmt='sqlite', saves to a SpatiaLite database.

returnxy

Logical. If TRUE, save xy coordinates to outfolder.

exportsp

Logical. If returnxy=TRUE, if TRUE, saves xy data as spatial data. If FALSE, saves xy data as table.

savedata_opts

List. See help(savedata_options()) for a list of options. Only used when savedata = TRUE.

spXYdat

R list object. Output from spGetXY().

gui

Logical. If TRUE, uses gui interface.

Value

xypltx

sf object. Input xy data clipped to boundary.

bndx

sf object. Input bnd.

tabs

list object. List of input layers clipped to boundary (pltx,condx,etc.).

xy.uniqueid

String. Name of unique identifier of xy.

puniqueid

String. Name of unique identifier of plot in plt.

pjoinid

String. Name of unique identifier of plot in plt.

If savedata=TRUE, outdat data frame is saved to outfolder.

Details

datsource

Plots are extracted from 3 different data sources:
1) CSV - data have previously been extracted from the FIA database and stored as CSV files.
2) datamart - data are extracted from FIA's publically-available datamart.
3) sqlite - data have previously been extracted from the FIA database and stored within a SQLite database.

Selection parameters

FIA plots are selected based on the following parameters:

evalid - the FIA evaluation identifier

evalCur - the most current FIA evaluation in database

evalEndyr - the FIA evaluation ending in evalEndyr

evalType - the FIA evaluation type ('ALL', 'AREAVOL', 'GRM', 'P2VEG', 'DWM', 'INV', 'REGEN', 'CRWN')

measCur - the most current measurement of each plot in database

measEndyr - the most current measuremtn of each plot in database in or prior to measEndyr

Endyr.filter - a filter for bnd that specifies the boundary where measEndyr should be applied

Note

If savebnd=TRUE:
If out_fmt=c('csv','shp'), the writeOGR (rgdal) function is called. The ArcGIS driver truncates variable names to 10 characters or less. Variable names are changed before export using an internal function (trunc10shp). If Spatial object has more than 1 record, it will be returned but not exported.

If datsource="datmart", data are imported from FIA DataMart. The plot coordinates have been altered for privacy (See https://www.fia.fs.fed.us/tools-data/spatial/Policy/index.php for details). The zip files are extracted on-the-fly from the online website. Web server connections will affect download speeds.

Author

Tracey S. Frescino

Examples

if (FALSE) {
# Get polygon vector layer from FIESTA external data
WYbhfn <- system.file("extdata",
                      "sp_data/WYbighorn_adminbnd.shp",
                      package = "FIESTA")

# Extract data from FIA datamart for measurement years 2013 thru 2015
dat <- spGetPlots(bnd = WYbhfn,
                  datsource = "datamart",
                  measyrs = 2013:2015)
names(dat)
tabs <- dat$tabs
names(tabs)
head(tabs$pltx)

table(tabs$pltx$MEASYEAR)

# Extract data from FIA datamart for most current evaluation
datCur <- spGetPlots(bnd = WYbhfn,
                     datsource = "datamart",
                     evalCur = TRUE)
names(datCur)
tabsCur <- datCur$tabs
names(tabsCur)
head(tabsCur$pltx)

table(tabsCur$pltx$MEASYEAR)
}